// site/src/types/routes.ts
export enum Routes {
  HOME = 'home',
  ABOUT = 'about',
  DETAILS = 'details',
  KYIV_GENEVA = 'kyiv-geneva',
  HOROWITZ = 'horowitz',
  COMPETITION = 'competition',
  PROJECTS = 'projects',
  MASTER_CLASS = 'master-class',
  SUMMER_SCHOOL = 'summer-school',
  VIRTUOSES = 'virtuoses',
  NEWS = 'news',
  CONTACTS = 'contacts',
}